Admin module, Custom error handling and Event CRUD

+++ BusinessLayer/EventRepository.cs
@@ -12,38 +12,44 @@ namespace BusinessLayer
     public class EventRepository
     {
         private AmazingRaceDBContext context = new AmazingRaceDBContext();
-            public EventRepository()
-            {
+        public EventRepository()
+        {
 
-            }
+        }
 
-            public IList<Event> GetAll()
-            {
-                return context.Events.ToList();
-            }
+        public IList<Event> GetAll()
+        {
+            return context.Events.ToList();
+        }
 
-            public void Add(Event entity)
+        public void Add(Event entity)
+        {
+            if (context != null && context.Events != null)
             {
                 context.Events.Add(entity);
                 context.SaveChanges();
             }
+        }
 
-            public void Remove(Event entity)
+        public void Remove(Guid id)
+        {
+            if (context != null && context.Events != null)
             {
-                var obj = context.Events.Find(entity.Id);
+                var obj = context.Events.Find(id);
                 context.Events.Remove(obj);
                 context.SaveChanges();
             }
+        }
 
-            public Event GetById(Guid id)
-            {
-                return context.Events.Find(id);
-            }
+        public Event GetById(Guid id)
+        {
+            return context.Events.Find(id);
+        }
 
-        public void update (Event eventToUpdate)
+        public void update(Event eventToUpdate)
         {
             context.Entry(eventToUpdate).State = EntityState.Modified;
             context.SaveChanges();
         }
     }
-    }
+}
